Separate into source files
also added a constructor
